Add completion to ParticipationByCountryBlock, use canvas implementation and improve tooltip
