+++ config/block_templates.yml
@@ -48,26 +48,26 @@ recommendedResourcesTemplate:
 # Used for salaries, company size, experience, etc. blocks that only show the current year of data
 demographicsTemplate:
     blockType: VerticalBarBlock
-    dataPath: 'surveyApi.survey.demographics.${id}.year' # select the 4th array item to select the 4th year
+    dataPath: 'surveyApi.survey.demographics.${id}.year'
     showDescription: true
     query: >
         surveyApi {
             survey(survey: ${surveyType}, year: ${currentYear}) {
                 demographics {
                     ${id} {
-                      year(year: ${currentYear}) {
-                        year
-                        total
-                        completion {
-                            percentage
-                            count
+                        year(year: ${currentYear}) {
+                            year
+                            total
+                            completion {
+                                percentage
+                                count
+                            }
+                            buckets {
+                                id
+                                count
+                                percentage
+                            }
                         }
-                        buckets {
-                            id
-                            count
-                            percentage
-                        }
-                      }
                     }
                 }
             }
