+++ src/core/blocks/demographics/ParticipationByCountryBlock/ParticipationByCountryBlock.js
@@ -2,13 +2,13 @@ import React, { memo, useState } from 'react'
 import PropTypes from 'prop-types'
 import Block from 'core/blocks/block/Block'
 import ChartContainer from 'core/charts/ChartContainer'
-import ParticipationByCountryChart from 'core/charts/demographics/ParticipationByCountryChart'
+import ParticipationByCountryChart from './ParticipationByCountryChart'
 
 const ParticipationByCountryBlock = ({ block, data, units: defaultUnits = 'percentage' }) => {
     const [units, setUnits] = useState(defaultUnits)
 
     return (
-        <Block units={units} setUnits={setUnits} data={data.buckets} block={block}>
+        <Block units={units} setUnits={setUnits} data={data} block={block}>
             <ChartContainer height={500}>
                 <ParticipationByCountryChart units={units} data={data.buckets} />
             </ChartContainer>
@@ -21,9 +21,15 @@ ParticipationByCountryBlock.propTypes = {
         id: PropTypes.string.isRequired
     }).isRequired,
     data: PropTypes.shape({
+        completion: PropTypes.shape({
+            count: PropTypes.number.isRequired,
+            percentage: PropTypes.number.isRequired
+        }).isRequired,
         buckets: PropTypes.arrayOf(
             PropTypes.shape({
-                id: PropTypes.string.isRequired
+                id: PropTypes.string.isRequired,
+                count: PropTypes.number.isRequired,
+                percentage: PropTypes.number.isRequired
             })
         ).isRequired
     }).isRequired
