+++ src/core/blocks/demographics/ParticipationByCountryBlock/index.js
@@ -1,0 +1,3 @@
+import ParticipationByCountryBlock from './ParticipationByCountryBlock'
+
+export default ParticipationByCountryBlock
